cleanup: moved files to specific folders
