+++ src/components/Navbar.jsx
@@ -1,5 +1,4 @@
 import { Cross as Hamburger } from "hamburger-react";
-import { FaBehance,FaLinkedinIn,FaXTwitter,FaInstagram } from "react-icons/fa6";
 import { useState } from "react";
 import { Link } from "react-router-dom";
 import { motion } from "framer-motion";
